Fix Split Two Columns folder name case
We should use camel case
